Fix a bug in MathsUtil (wut??), start laying groundwork for taking into account segment area (but not 100% happy with it yet)

+++ Core/src/burlton/core/code/util/MathsUtil.kt
@@ -18,8 +18,7 @@ class MathsUtil
             return if (count == 0)
             {
                 0.0
-            } else round(100 * count / total, 1)
-
+            } else round(100 * count.toDouble() / total, 1)
         }
     }
 }
